Move build_transport_info to pipeline
